podman: Use sudo to execute podman command
Since rootless support is not there for minc, and we don't want user
to run the minc with sudo by default so we are using `sudo podman`
command on linux to make sure only those command executed in sudo mode.
This will also fix the kubeconfig file which is created on `root` user's
home instead user who is running it.

Signed-off-by: Praveen Kumar <kumarpraveen.nitdgp@gmail.com>